+++ minecraft/net/minecraft/src/ItemMarteau.java
@@ -1,0 +1,16 @@
+package net.minecraft.src;
+
+public class ItemMarteau extends Item
+{
+
+	protected ItemMarteau(int par1)
+	{
+		super(par1);
+	}
+	
+	public boolean isFull3D()
+	{
+		return true;
+	}
+
+}